# shared-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: VibeCare API
  version: 0.7.0
  description: Canonical mobile API. FITRUS credentials never cross this boundary.
servers:
  - url: https://api.example.invalid
    description: Replace per environment
security:
  - bearerAuth: []
paths:
...
  /v1/algorithm-rules/current:
    get:
      responses:
        '200': { description: Active validated AlgorithmRuleSet }
        '503': { description: No enabled valid rule set; execution is disabled }
  /v1/fitrus/measurements/{kind}:
    post:
      description: Server-only FITRUS proxy. The payload must match the selected provider endpoint contract. Valid body-composition and vital responses are inserted into history; unsupported or model-error responses remain raw only.
      parameters:
        - in: path
          name: kind
          required: true
          schema:
            enum: [bodyFat, bloodPressure, heartRate, stress, stressV2, bodyTemperature]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deviceId, payload]
              properties:
                deviceId: { type: string }
                measuredAt: { type: string, format: date-time }
                muscleProvenance:
                  description: Required when kind=bodyFat. UNKNOWN, empty references, or a method/evidence/definition triple absent from the active rule's applicableMethods cannot become simulation-eligible.
                  type: object
                  required: [muscleDefinition, definitionRef, muscleMeasurementMethod, methodEvidenceRef, muscleMassUnit, acquisitionProtocol]
                  properties:
                    muscleDefinition: { type: string, enum: [UNKNOWN, ASM, SMM] }
                    definitionRef: { type: string, minLength: 1 }
                    muscleMeasurementMethod: { type: string, minLength: 1 }
                    methodEvidenceRef: { type: string, minLength: 1 }
                    muscleMassUnit: { const: kg }
                    acquisitionProtocol: { type: string, minLength: 1 }
                payload:
                  description: Must match the schema for kind. PPG list length/range remains device-spec dependent.
                  oneOf:
                    - { $ref: '#/components/schemas/FitrusBloodPressureRequest' }
                    - { $ref: '#/components/schemas/FitrusHeartRateRequest' }
                    - { $ref: '#/components/schemas/FitrusStressRequest' }
                    - { $ref: '#/components/schemas/FitrusStressV2Request' }
                    - { $ref: '#/components/schemas/FitrusBodyTemperatureRequest' }
                    - { $ref: '#/components/schemas/FitrusBodyFatRequest' }
      responses:
        '201': { description: Raw response retained. normalized indicates whether a validated history record was also inserted. }
        '400': { description: Invalid wrapper or endpoint-specific FITRUS payload }
        '409': { description: Body-fat demographics do not match the authenticated participant profile }
        '502': { description: FITRUS rejected the request or could not be reached }
        '504': { description: FITRUS did not respond before the server timeout }
  /v1/participants/me/measurement-set/current:
    get:
      description: Returns the latest 20 history records plus any older records in the selected four (at most 24). Selection uses the requested source device, measuredAt descending, then ID descending.
      parameters:
        - in: query
          name: deviceId
          schema: { type: string }
      responses:
        '200': { description: Current body-composition set }
...
  /v1/recommendations/authorize:
    post:
      description: pilot-0.9.0 accepts four provenance-matched SMM measurements and uses the latest height-normalized skeletal-muscle level with the selected part's fixed time, frequency, and intensity. Physical execution is always PROHIBITED.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [measurementIds, safety, deviceId, algorithmVersion, bodyPart]
              properties:
                measurementIds:
                  type: array
                  minItems: 4
                  maxItems: 4
                  uniqueItems: true
                  items: { type: string }
                safety: { $ref: '#/components/schemas/SafetyCheck' }
                deviceId: { type: string }
                sourceDeviceId:
                  type: string
                  minLength: 1
                  description: BIA measurement source device. Legacy callers may omit it to use deviceId. deviceId identifies the target simulator.
                algorithmVersion: { type: string }
                bodyPart:
                  type: string
                  enum: [wholeBody, shoulder, arm, abdomen, thigh, calf]
                  description: Selects the configured time, frequency, and baseline device-output setting. All four measurements must have confirmed SMM provenance.
                requestedIntensityPct:
                  type: integer
                  minimum: 1
                  maximum: 100
                  description: Optional device output setting in percent of the device's own output scale. The server accepts only values between its configured minimum and calculated recommendation. This is not a measured physical output.
      responses:
        '201':
          description: One-time 60-second mock authorization.
          content:
            application/json:
              schema:
                type: object
                required: [authorized, mode, authorizationId, recommendationId, expiresAt, result]
                properties:
                  authorized: { const: true }
                  mode: { const: mock }
                  authorizationId: { type: string }
                  recommendationId: { type: string }
                  expiresAt: { type: string, format: date-time }
                  result: { $ref: '#/components/schemas/AlgorithmResult' }
        '409': { description: 'Non-SMM or mixed provenance, stale/future data under the labelled engineering policy, safety block, physical mode, intensity range, or algorithm version mismatch.' }
        '400': { description: Invalid input or missing explicit safety answers }
        '503': { description: Algorithm unavailable }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    FitrusPpgList:
      type: array
      minItems: 1
      items: { type: integer }
      description: Interleaved red/green/IR integers. The provider has not supplied a valid sample count or item range.
    FitrusBloodPressureRequest:
      type: object
      required: [list, baseSystolic, baseDiastolic]
      properties:
        list: { $ref: '#/components/schemas/FitrusPpgList' }
        baseSystolic: { type: number, exclusiveMinimum: 0 }
        baseDiastolic: { type: number, exclusiveMinimum: 0 }
    FitrusHeartRateRequest:
      type: object
      required: [list]
      properties:
        list: { $ref: '#/components/schemas/FitrusPpgList' }
    FitrusStressRequest:
      type: object
      required: [list, age]
      properties:
        list: { $ref: '#/components/schemas/FitrusPpgList' }
        age: { type: integer, minimum: 1 }
    FitrusStressV2Request:
      $ref: '#/components/schemas/FitrusHeartRateRequest'
    FitrusBodyTemperatureRequest:
      type: object
      required: [temp]
      properties:
        temp: { type: number }
    FitrusBodyFatRequest:
      type: object
      required: [age, height, weight, gender, voltage]
      properties:
        age: { type: integer, minimum: 1 }
        height: { type: number, exclusiveMinimum: 0 }
        weight: { type: number, exclusiveMinimum: 0 }
        gender: { type: string, enum: [male, female], description: Case-insensitive on input }
        voltage: { type: number }
        correct: { type: [number, 'null'], default: 0 }
    AlgorithmResult:
      type: object
      required: [status, executionStatus, dataDecision, simulationEligibility, physicalExecution, realDeviceSendAllowed, reasonCodes, warnings, algorithmVersion, bodyPart, muscleMassBasis, average, recommendation, factors]
      properties:
        status: { type: string, enum: [READY, REVIEW, BLOCKED] }
        executionStatus: { type: string, enum: [SIMULATION_READY, REVIEW, BLOCKED] }
        dataDecision: { type: string, enum: [ACCEPTED, REVIEW_REQUIRED, BLOCKED] }
        simulationEligibility: { type: string, enum: [ELIGIBLE, INELIGIBLE] }
        physicalExecution: { const: PROHIBITED }
        realDeviceSendAllowed: { const: false }
        algorithmVersion: { const: pilot-0.9.0 }
        bodyPart: { type: string, enum: [wholeBody, shoulder, arm, abdomen, thigh, calf] }
        muscleMassBasis: { const: SMM }
        reasonCodes: { type: array, items: { type: string } }
        warnings: { type: array, items: { type: string } }
        average:
          oneOf:
            - type: 'null'
            - type: object
              required: [weightKg, bmi, bodyFatPct, fatMassKg, skeletalMuscleMassKg]
              properties:
                weightKg: { type: number }
                bmi: { type: number }
                bodyFatPct: { type: number }
                fatMassKg: { type: number }
                skeletalMuscleMassKg: { type: number }
        recommendation:
          oneOf:
            - type: 'null'
            - type: object
              required: [purpose, evidence, durationSec, frequencyHz, intensityPct, baseIntensityPct]
              properties:
                purpose: { const: SIMULATION_CANDIDATE }
                evidence: { const: HYPOTHESIS_UNVALIDATED }
                durationSec: { type: integer, minimum: 1 }
                frequencyHz: { type: integer, minimum: 1 }
                intensityPct: { type: number, minimum: 0, maximum: 100 }
                baseIntensityPct: { type: number, minimum: 0, maximum: 100 }
        factors:
          oneOf:
            - type: 'null'
            - type: object
              required: [genderCoefficient, ageCoefficient, bodyFatCoefficient, muscleMassCoefficient, totalCoefficient, calculatedIntensityPct, muscleLevel, muscleIndexKgM2]
              properties:
                genderCoefficient: { type: number, exclusiveMinimum: 0, maximum: 1 }
                ageCoefficient: { type: number, exclusiveMinimum: 0, maximum: 1 }
                bodyFatCoefficient: { type: number, exclusiveMinimum: 0, maximum: 1 }
                muscleMassCoefficient: { const: 1 }
                totalCoefficient: { type: number, exclusiveMinimum: 0, maximum: 1 }
                calculatedIntensityPct: { type: number, minimum: 0, maximum: 100 }
                muscleLevel: { type: string, enum: [low, medium, high] }
                muscleIndexKgM2: { type: number, exclusiveMinimum: 0 }
    SafetyCheck:
      description: All three answers must be explicitly provided; null or omitted is not false. Any true answer blocks execution.
      type: object
      required: [acutePain, dizziness, clinicianHold]
      properties:
        acutePain: { type: boolean }
        dizziness: { type: boolean }
        clinicianHold: { type: boolean }